Allow any custom tag to be set
We want to pick a custom image tag to use, i.e. `dev`. Others are likely
to want to pick `latest`, while someone else might need `main`. This
allows any custom image tag to be set, and it will be added to the
generated versions.

While this was the simplest thing, we may want to add support for
multiple tags in the future.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>